Nearly finished with UI
The prompt is shown every time a new page is loaded, so that
it is easy to test the UI. Currently it does not reset occasionally,
for a reason that I have been unable to pin down. After interacting with
the panel and closing it, the panel will sometimes not "reset" itself
for the next interaction.